adjust readme, add some more functionalities

+++ src/IndusLinkApp/src/main/java/com/induslink/app/service/entities/DeviceService.java
@@ -2,6 +2,7 @@ package com.induslink.app.service.entities;
 
 import com.induslink.app.models.entities.Device;
 import com.induslink.app.repository.entities.DeviceRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -19,7 +20,8 @@ public class DeviceService {
     }
 
     public Optional<Device> getDeviceById(Long id) {
-        return deviceRepository.findById(id);
+        return Optional.ofNullable(deviceRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Device with ID " + id + " not found")));
     }
 
     public Device saveDevice(Device device) {
@@ -27,6 +29,10 @@ public class DeviceService {
     }
 
     public void deleteDevice(Long id) {
+        if (!deviceRepository.existsById(id)) {
+            throw new EntityNotFoundException("Device with ID " + id + " not found");
+        }
         deviceRepository.deleteById(id);
     }
 }
+
